split `TurnObjManager`
